Moving param to types folder common
