+++ packages/common/src/index.ts
@@ -1,4 +1,3 @@
 export * from './validation';
 export * from './types';
 export * from './schema';
-export * from './params';
