--- packages/common/src/params/index.ts
@@ -1,1 +1,0 @@
-export * from './group';
